+++ sql/scripts/create-mysql.sql
@@ -1,26 +1,26 @@
 
 -- This is an example table create script for MySQL. It is not required for running the unit tests, which use H2.
 
-create table vertexium_vertex (
-  id varchar(255) primary key,
-  object longtext not null
+create table visallo_vertex (
+  id varchar(767) primary key,
+  object longblob not null
 );
 
-create table vertexium_edge (
-  id varchar(255) primary key,
-  in_vertex_id varchar(255),
-  out_vertex_id varchar(255),
-  object longtext not null
+create table visallo_edge (
+  id varchar(767) primary key,
+  in_vertex_id varchar(767),
+  out_vertex_id varchar(767),
+  object longblob not null
 );
 
-create table vertexium_metadata (
-  id varchar(255) primary key,
-  object longtext not null
+create table visallo_metadata (
+  id varchar(767) primary key,
+  object longblob not null
 );
 
-create table vertexium_streaming_properties (
-  id varchar(255) primary key,
+create table visallo_streaming_properties (
+  id varchar(767) primary key,
   data longblob not null,
-  type varchar(255) not null,
+  type varchar(4000) not null,
   length bigint not null
 );